% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Title
 This is a big mess}
\usage{
f(x, y)
}
\arguments{
\item{x}{This is the first argument}

\item{y}{This is the second argument}
}
\value{
Something I hope
}
\description{
Title
 This is a big mess
}
\examples{
f(2,3)

}
